No longer act as a web server
Having the Pi Pico W operate as a web server and responding to pull
requests for sensor data once a minute turned out to be quite expensive
in terms of CPU power. A set of AA batteries powering the board only
kept going for 24 hours or so.

Instead I'm turning the idea around, and the board will spend most of
its time asleep. Properly asleep that is, rather than constantly
listening for an incoming HTTP request.

Periodically (initially set to 5 minutes) it will wake up, query the
sensors on the board, and send the data to a server that is sitting
inside my house, powered by mains power! That's the system I'll be using
to analyse the data and turn it into time series plots anyway.

This commit removes the HTTP server functionality and puts in the
skeleton for the push, although that part needs to be finished off.